move useResolverToRef to separate file
