+++ src/hooks/accounts/resolvers/useGetAccountsQueryResolver.tsx
@@ -2,7 +2,7 @@ import { isArray } from 'lodash';
 import { useCallback } from 'react';
 import { useGetAccounts } from '../useGetAccounts';
 import { usePersistActiveAccount } from '../usePersistActiveAccount';
-import { useResolverToRef } from './useAccountsMutationResolvers';
+import { useResolverToRef } from '../../apollo/useResolverToRef';
 
 export interface AccountsQueryResolverArgs {
     isActive?: boolean
