+++ src/hooks/accounts/resolvers/useSetActiveAccountMutationResolver.tsx
@@ -1,12 +1,12 @@
-import { ApolloCache, ApolloClient, NormalizedCacheObject } from '@apollo/client';
+import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { useCallback } from 'react';
 import { useSetActiveAccount } from '../useSetActiveAccount'
 import { SetActiveAccountMutationVariables } from '../mutations/useSetActiveAccountMutation';
-import { useResolverToRef } from './useAccountsMutationResolvers';
+import { useResolverToRef } from '../../apollo/useResolverToRef';
 
 export const useSetActiveAccountMutationResolver = () => {
     const setActiveAccount = useSetActiveAccount();
-    
+
     return useResolverToRef(
         useCallback(async (
             _obj,
