+++ src/hooks/extension/resolvers/query/extension.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { Extension } from '../../../../generated/graphql';
-import { useResolverToRef } from '../../../accounts/resolvers/useAccountsMutationResolvers';
+import { useResolverToRef } from '../../../apollo/useResolverToRef';
 import { getExtension } from '../../lib/getExtension';
 
 // make sure the __typename is well typed
