+++ src/hooks/pools/resolvers/useSubmitTradeMutationResolvers.tsx
@@ -4,7 +4,7 @@ import { gql } from 'graphql.macro';
 import { useCallback } from 'react'
 import { PoolType } from '../../../components/Chart/shared';
 import { Maybe, TradeType } from '../../../generated/graphql';
-import { useResolverToRef } from '../../accounts/resolvers/useAccountsMutationResolvers'
+import { useResolverToRef } from '../../apollo/useResolverToRef'
 import { usePolkadotJsContext } from '../../polkadotJs/usePolkadotJs';
 import { SubmitTradeMutationVariables } from '../mutations/useSubmitTradeMutation';
 import { buy as buyLbp } from '../lbp/buy';
