+++ src/hooks/vesting/useVestingScheduleQueryResolvers.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react'
 import { Account } from '../../generated/graphql';
-import { useResolverToRef } from '../accounts/resolvers/useAccountsMutationResolvers'
+import { useResolverToRef } from '../apollo/useResolverToRef'
 import { useGetVestingScheduleByAddress } from './useGetVestingScheduleByAddress';
 
 export const useVestingScheduleQueryResolvers = () => {
